split hook
